[Docs] Change folder names
기존 experiment폴더를 model로 변경
issue #64
