+++ model/imagecluster/calc.py
