+++ model/laughter/laughter_detector/models.py
